web/backend-queue-new-data/requirements.txt
alembic==1.0.5
apturl==0.5.2
asn1crypto==0.24.0
autopep8==1.4.3
Brlapi==0.6.6
catfish==1.4.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
httplib2==0.9.2
idna==2.7
iexfinance==0.3.5
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.0
menulibre==2.2.0
mugshot==0.4.0
numpy==1.15.4
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
onboard==1.4.1
pandas==0.23.4
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
PySocks==1.6.8
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
python-debianbts==2.7.2
python-editor==1.0.3
python-magic==0.4.16
pytz==2018.7
pyxdg==0.25
PyYAML==3.13
reportlab==3.4.0
requests==2.20.1
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
searchtweets==1.7.4
SecretStorage==2.3.1
sgt-launcher==0.2.4
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.14
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tweepy==3.7.0
tweet-parser==1.13.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==0.14.1
xcffib==0.5.1
xkit==0.0.0
zope.interface==4.3.2
